\name{beta.limit}
\alias{beta.limit}
\docType{data}
\title{
Upper limit of different beta diversity (dissimilarity) indexes
}
\description{
Upper limit value of each abundance-based or incidence-based dissimilarity index.
}
\usage{data("beta.limit")}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{Dmax.in}}{numeric, upper limit of incidence-based dissimilarity}
    \item{\code{Dmax.ab}}{numeric, upper limit of abundance-based dissimilarity}
  }
}
\examples{
data(beta.limit)
}
\keyword{indexes}
\keyword{datasets}
